Remove wrong test
This test contradicts the task. This is not runtime error.
But it totally fits the definition of Syntax error. It is impossible to deduce this string from the grammar.
Futhermore, other runtime error tests in this test case actually not list-tests. They fit only <application> symbol definition. It really IS runtime error though.

Implementation of scheme from MIT (mit-scheme) says it is valid expression that produces "()". Implementation scheme48 says: "syntax-violation: invalid expression".